cmd/derper: move 204 handler from package main to derphttp
Updates #13038
